refactor(config): isolate normalization and travel defaults
